pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "clawforge"
version = "1.0.0"
description = "Institutional 5M crypto trading bot for BingX"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "Proprietary"}
authors = [
    {name = "Zion", email = "zion@clawforge.ai"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Financial and Insurance Industry",
    "License :: Other/Proprietary License",
    "Programming Language :: Python :: 3.11",
    "Topic :: Office/Business :: Financial :: Investment",
]
dependencies = [
    "freqtrade>=2024.1",
    "pandas>=2.0.0",
    "pandas-ta>=0.3.14b0",
    "ccxt>=4.0.0",
    "python-telegram-bot>=20.0",
    "stepfun>=1.0.0",
    "Pillow>=10.0.0",
    "python-dotenv>=1.0.0",
    "sqlalchemy>=2.0.0",
    "requests>=2.31.0",
    "aiohttp>=3.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
    "pre-commit>=3.0.0",
    "types-requests>=2.31.0",
]

[tool.setuptools]
packages = ["clawforge"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
target-version = "py311"
line-length = 100
select = ["E", "F", "B", "I", "N", "UP", "YTT", "S", "BLE", "A", "C4", "DTZ", "T10", "EM", "EXE", "ISC", "ICN", "G", "LOG", "PIE", "PT", "Q", "RSE", "RET", "SLF", "SLOT", "TCH", "INT", "PTH", "FBT", "ASYNC", "FLY", "PERF", "RUF"]
ignore = ["S101", "S603", "S607"]

[tool.ruff.per-file-ignores]
"tests/*" = ["S101", "PLR2004"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
strict_optional = true

[[tool.mypy.overrides]]
module = ["freqtrade.*", "ccxt.*", "telegram.*", "pandas_ta.*"]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "--cov=clawforge --cov-report=term-missing --cov-report=html --strict-markers"

[tool.coverage.run]
source = ["clawforge"]
omit = ["*/tests/*", "*/venv/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
]
